chore: 타입 변경

+++ mashup-member/src/main/java/kr/mashup/branding/ui/schedule/response/ScheduleResponseListWithProgress.java
@@ -11,7 +11,7 @@ public class ScheduleResponseListWithProgress {
 
     Progress progress;
 
-    Long dateCount;
+    Integer dateCount;
 
     List<ScheduleResponse> scheduleList;
 }
